refactor: cleanup unused code and fix inconsistencies
- Rename dificulty-badge.tsx to difficulty-badge.tsx (typo fix)
- Update all imports referencing the renamed file
- Remove unused ArrayElement type from lib/utils.ts
- Remove unused _timingMiddleware from server/api/trpc.ts
- Standardize logging: replace console.log/error with Sentry logger
  in lib/resend.ts and app/api/cli/challenge/[slug]/submit/route.ts
