+++ components/difficulty-badge.tsx
