+++ lib/utils.ts
@@ -4,5 +4,3 @@ import { twMerge } from "tailwind-merge";
 export function cn(...inputs: ClassValue[]) {
   return twMerge(clsx(inputs));
 }
-
-export type ArrayElement<T> = T extends readonly (infer U)[] ? U : never;
